Add files via upload
Process_data_V3.m ==> read the joints data from bag file and cut the parts that the robot was steady in the beginning and the end. The Time, joint position, velocity, and acceleration are saved to mat file in the same directory as the bag file with the same name.